requirements.txt
pyTelegramBotAPI
openai
python-dotenv
SpeechRecognition
pyttsx3
httpx[http2,socks]